rack selected and item dispensed if coins added are sufficient

+++ src/main/java/vendables/Crisps.java
@@ -2,7 +2,7 @@ package vendables;
 
 public class Crisps extends VendableItem implements IVend {
 
-    public Crisps(String name, double price){
+    public Crisps(String name, int price){
         super(name, price);
     }
