+++ src/main/java/vendables/Drink.java
@@ -4,7 +4,7 @@ public class Drink extends VendableItem implements IVend {
 
     private int volume;
 
-    public Drink(String name, double price, int volume) {
+    public Drink(String name, int price, int volume) {
         super(name, price);
         this.volume = volume;
     }
