+++ src/main/java/vendables/Sweet.java
@@ -2,7 +2,7 @@ package vendables;
 
 public class Sweet extends VendableItem implements IVend {
 
-    public Sweet(String name, double price){
+    public Sweet(String name, int price){
         super(name, price);
     }
 }
